backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
groq
pytesseract
pdf2image
python-docx
reportlab
python-dotenv
pypdf
